Minor changes to time keeping, added Squid Blind
Did some minor change (optimisation?) to how timing was kept by moving
when the timing scores are reset into the functions that execute based
on them.
Added Squid Blind, now when a squid is damaged in a 1 second interval it
will give the Blindness effect to any enities within 4 blocks.