Removed Horse/Donkey meat
Horses/Donkies can drop "Raw Horse/Donkey" which is just renamed Mutton,
and if on fire they can drop "Cooked Horse/Donkey", however if the Raw
meat is cooked in a Furance it is currently not possible to rename that
to Horse/Donkey meat. As such I have decided to place this into an
unused folder until such a time as Furnaces have the functionality to
rename items, or custom items can be added.